refatorada a classe cliente para inclusão do método str, inseridos métodos de listagem na classe mercado, refatorado o arquivo main para rodar o programa
