+++ .gitignore
@@ -1,0 +1,7 @@
+
+__pycache__/
+*.py[cod]
+
+*.pyo
+*.pyc
+*.pyd
